requirements.txt
streamlit
numpy
opencv-python
Pillow
transformers
torch
torchvision
scikit-image